Se corrigen errores

+++ backend/src/application/use_cases/TableUseCase/DeleteTable.ts
@@ -1,4 +1,5 @@
 import { TableRepository } from "../../../infrastructure/database/repository/TableRepository.js";
+import { BusinessError } from "../../../shared/exceptions/BusinessError.js";
 import { NotFoundError } from "../../../shared/exceptions/NotFoundError.js";
 
 export class DeleteTable {
@@ -6,10 +7,12 @@ export class DeleteTable {
         private readonly tableRepository = new TableRepository
     ) {}
 
-    public async execute (numTable : number): Promise<void> {
+    public async execute (numTable: number): Promise<void> {
         const tableFound = await this.tableRepository.getByNumTable(numTable);
         
         if (!tableFound) throw new NotFoundError("Mesa no encontrada");
+
+        if(tableFound.state == 'Ocupada') throw new BusinessError('No puedes eliminar una mesa que se encuntra ocupada.')
 
         await this.tableRepository.deleteTable(numTable);
     }
